clean up mapping
